Add Consultation model (#9)
This changes some of the basic structures of the codebase. Now, the Question
model is basically invisible to the user. What he sees is the Consultation's
question, each Earl name (as category), and the Choices (that seems related to
Earl, but are related to its Question).

The user, instead of creating Questions, creates Consultations and Categories
(Earls). The idea is that there's a Consultation "How to fix the world?", then
the Choices might be in a few categories (Earl). For example, in Healthcare,
one choice could be "End all diseases".