Create or sign in as AOI user when coming from Facebook (#10)
When our app is viewed inside a Facebook Canvas, FB sends a signed_request
parameter. We use it to gather facebook's user_id and e-mail. With it, we
create a new "local" user and sign in as that.

After it, it's transparent for our app if the user is from facebook or not.