Ties facebook authentication with our normal auth method (closes #10)
There're three parts in this commit.

First, we tie the normal and FB's authentication methods together, so
authenticate works doesn't matter if it's coming from FB or not.

Second, we try to authenticate ONLY through facebook whenever you go to the
"New Consultation" page. I couldn't use the aforementioned authenticate method
because, if the user isn't coming through FB, this method will redirect her to
the sign in page. As she doesn't have an user yet, she would be stuck.

The last part is logging in the user in our system if she's coming from FB.
When you access our page through Facebook Canvas, it sends a POST request with
a signed_request param. This is only sent to the first page that the user
access. As she might access any page, I had to check it on a before filter
in ApplicationController.