Detect whitelabel using the URL, not the session.
Before, to set whitelabel for the Facebook App, we checked if it was a Facebook
request and saved that value to the session. After that, we would use the
session.

The problem with that approach is that, if the same user visits the site from
outside Facebook, the cookie will still be set and she'll see the whitelabel
version. For ever.

We can't simply check if it's a Facebook request every time because FB only
sends us something in the first page that the user hits. After that, there's no
way (AFAIK) for us to know if the user is vising us through Facebook or not.

To solve this, instead of saving in the session, we check the URL that the user
is coming from. If it's something like "whitelabel.allourideas.org", we show
the whitelabel; if it's "allourideas.org", we don't.

It's the same strategy adopted by the widget page.